All deprecated code is removed. Java package changed to match maven group and artifact id.
