+++ src/main/java/com/github/cliftonlabs/json_simple/Yylex.java
@@ -1,16 +1,14 @@
-/* The following code was generated by JFlex 1.4.3 on 8/22/16 3:26 PM */
+/* The following code was generated by JFlex 1.4.3 on 1/9/18 6:59 PM */
 
-package org.json.simple.parser;
+package com.github.cliftonlabs.json_simple;
 
 
 /**
  * This class is a scanner generated by 
  * <a href="http://www.jflex.de/">JFlex</a> 1.4.3
- * on 8/22/16 3:26 PM from the specification file
- * <tt>/home/davinloegering/cliftonlabs/workspace/json-simple/src/main/lex/json.lex</tt>
- * 
+ * on 1/9/18 6:59 PM from the specification file
+ * <tt>/home/loegering/cliftonlabs/workspace/json-simple/src/main/lex/jsonstrict.lex</tt>
  */
-@Deprecated
 class Yylex {
 
   /** This character denotes the end of file */
@@ -57,8 +55,8 @@ class Yylex {
     "\2\0\2\1\1\2\1\3\1\4\3\1\1\5\1\6"+
     "\1\7\1\10\1\11\1\12\1\13\1\14\1\15\5\0"+
     "\1\14\1\16\1\17\1\20\1\21\1\22\1\23\1\24"+
-    "\1\0\1\25\1\0\1\25\4\0\1\26\1\27\2\0"+
-    "\1\30";
+    "\1\0\1\2\1\0\1\2\4\0\1\25\1\26\2\0"+
+    "\1\27";
 
   private static int [] zzUnpackAction() {
     int [] result = new int[45];
@@ -276,7 +274,7 @@ class Yylex {
   private boolean zzEOFDone;
 
   /* user code: */
-private StringBuffer sb=new StringBuffer();
+private StringBuilder sb=new StringBuilder();
 
 int getPosition(){
 	return yychar;
@@ -511,7 +509,7 @@ int getPosition(){
    * @return      the next token
    * @exception   java.io.IOException  if any I/O-Error occurs
    */
-  public Yytoken yylex() throws java.io.IOException, ParseException {
+  public Yytoken yylex() throws java.io.IOException, DeserializationException {
     int zzInput;
     int zzAction;
 
@@ -583,108 +581,104 @@ int getPosition(){
       zzMarkedPos = zzMarkedPosL;
 
       switch (zzAction < 0 ? zzAction : ZZ_ACTION[zzAction]) {
+        case 4: 
+          { sb = null; sb = new StringBuilder(); yybegin(STRING_BEGIN);
+          }
+        case 24: break;
         case 11: 
           { sb.append(yytext());
           }
         case 25: break;
+        case 5: 
+          { return new Yytoken(Yytoken.Types.LEFT_BRACE, null);
+          }
+        case 26: break;
         case 16: 
           { sb.append('\b');
           }
-        case 26: break;
-        case 6: 
-          { return new Yytoken(Yytoken.TYPE_RIGHT_BRACE,null);
-          }
         case 27: break;
         case 23: 
-          { Boolean val=Boolean.valueOf(yytext()); return new Yytoken(Yytoken.TYPE_VALUE, val);
+          { try{
+			int ch=Integer.parseInt(yytext().substring(2),16);
+			sb.append((char)ch);
+		}catch(Exception e){
+			/* The lexer is broken if it can build a 4 byte character code and fail to append the character. */
+			throw new DeserializationException(yychar, DeserializationException.Problems.UNEXPECTED_EXCEPTION, e);
+		}
           }
         case 28: break;
         case 22: 
-          { return new Yytoken(Yytoken.TYPE_VALUE, null);
+          { Boolean val=Boolean.valueOf(yytext()); return new Yytoken(Yytoken.Types.DATUM, val);
           }
         case 29: break;
-        case 13: 
-          { yybegin(YYINITIAL);return new Yytoken(Yytoken.TYPE_VALUE, sb.toString());
-          }
-        case 30: break;
         case 12: 
           { sb.append('\\');
           }
+        case 30: break;
+        case 10: 
+          { return new Yytoken(Yytoken.Types.COLON, null);
+          }
         case 31: break;
-        case 21: 
-          { Double val=Double.valueOf(yytext()); return new Yytoken(Yytoken.TYPE_VALUE, val);
+        case 9: 
+          { return new Yytoken(Yytoken.Types.COMMA, null);
           }
         case 32: break;
-        case 1: 
-          { throw new ParseException(yychar, ParseException.ERROR_UNEXPECTED_CHAR, new Character(yycharat(0)));
+        case 21: 
+          { return new Yytoken(Yytoken.Types.DATUM, null);
           }
         case 33: break;
-        case 8: 
-          { return new Yytoken(Yytoken.TYPE_RIGHT_SQUARE,null);
-          }
-        case 34: break;
         case 19: 
           { sb.append('\r');
           }
-        case 35: break;
+        case 34: break;
         case 15: 
           { sb.append('/');
           }
+        case 35: break;
+        case 2: 
+          { java.math.BigDecimal val= new java.math.BigDecimal(yytext()); return new Yytoken(Yytoken.Types.DATUM, val);
+          }
         case 36: break;
-        case 10: 
-          { return new Yytoken(Yytoken.TYPE_COLON,null);
-          }
-        case 37: break;
-        case 4: 
-          { sb = null; sb = new StringBuffer(); yybegin(STRING_BEGIN);
-          }
-        case 38: break;
         case 14: 
           { sb.append('"');
           }
-        case 39: break;
-        case 5: 
-          { return new Yytoken(Yytoken.TYPE_LEFT_BRACE,null);
+        case 37: break;
+        case 8: 
+          { return new Yytoken(Yytoken.Types.RIGHT_SQUARE, null);
           }
-        case 40: break;
+        case 38: break;
         case 17: 
           { sb.append('\f');
           }
-        case 41: break;
-        case 24: 
-          { try{
-														int ch=Integer.parseInt(yytext().substring(2),16);
-														sb.append((char)ch);
-													}
-													catch(Exception e){
-														throw new ParseException(yychar, ParseException.ERROR_UNEXPECTED_EXCEPTION, e);
-													}
+        case 39: break;
+        case 1: 
+          { throw new DeserializationException(yychar, DeserializationException.Problems.UNEXPECTED_CHARACTER, new Character(yycharat(0)));
           }
-        case 42: break;
+        case 40: break;
+        case 6: 
+          { return new Yytoken(Yytoken.Types.RIGHT_BRACE, null);
+          }
+        case 41: break;
         case 20: 
           { sb.append('\t');
           }
-        case 43: break;
+        case 42: break;
         case 7: 
-          { return new Yytoken(Yytoken.TYPE_LEFT_SQUARE,null);
+          { return new Yytoken(Yytoken.Types.LEFT_SQUARE, null);
           }
-        case 44: break;
-        case 2: 
-          { Long val=Long.valueOf(yytext()); return new Yytoken(Yytoken.TYPE_VALUE, val);
-          }
-        case 45: break;
+        case 43: break;
         case 18: 
           { sb.append('\n');
           }
-        case 46: break;
-        case 9: 
-          { return new Yytoken(Yytoken.TYPE_COMMA,null);
+        case 44: break;
+        case 13: 
+          { yybegin(YYINITIAL);return new Yytoken(Yytoken.Types.DATUM, sb.toString());
           }
-        case 47: break;
+        case 45: break;
         case 3: 
           { 
           }
-        case 48: break;
+        case 46: break;
         default: 
           if (zzInput == YYEOF && zzStartRead == zzCurrentPos) {
             zzAtEOF = true;
