--- src/main/java/org/json/simple/JSONObject.java
@@ -1,154 +1,0 @@
-/* Copyright 2006 FangYidong
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License. */
-package org.json.simple;
-
-import java.io.IOException;
-import java.io.StringWriter;
-import java.io.Writer;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-
-/**
- * A JSON object. Key value pairs are unordered. JSONObject supports java.util.Map interface.
- * 
- * @author FangYidong&lt;fangyidong@yahoo.com.cn&gt;
- * @deprecated since 2.0.0, replaced by {@link org.json.simple.JsonObject}
- */
-@Deprecated
-public class JSONObject extends HashMap implements Map, JSONAware, JSONStreamAware{
-	
-	private static final long serialVersionUID = -503443796854799292L;
-	
-	
-	/**
-	 * 
-	 */
-	public JSONObject() {
-		super();
-	}
-
-	/**
-	 * Allows creation of a JSONObject from a Map. After that, both the
-	 * generated JSONObject and the Map can be modified independently.
-	 * 
-	 * @param map description omitted.
-	 */
-	public JSONObject(Map map) {
-		super(map);
-	}
-
-
-    /**
-     * Encode a map into JSON text and write it to out.
-     * If this map is also a JSONAware or JSONStreamAware, JSONAware or JSONStreamAware specific behaviours will be ignored at this top level.
-     * 
-     * @see org.json.simple.JSONValue#writeJSONString(Object, Writer)
-     * 
-     * @param map description omitted.
-     * @param out description omitted.
-     * @throws IOException description omitted.
-     */
-	public static void writeJSONString(Map map, Writer out) throws IOException {
-		if(map == null){
-			out.write("null");
-			return;
-		}
-		
-		boolean first = true;
-		Iterator iter=map.entrySet().iterator();
-		
-        out.write('{');
-		while(iter.hasNext()){
-            if(first)
-                first = false;
-            else
-                out.write(',');
-			Map.Entry entry=(Map.Entry)iter.next();
-            out.write('\"');
-            out.write(escape(String.valueOf(entry.getKey())));
-            out.write('\"');
-            out.write(':');
-			JSONValue.writeJSONString(entry.getValue(), out);
-		}
-		out.write('}');
-	}
-
-	public void writeJSONString(Writer out) throws IOException{
-		writeJSONString(this, out);
-	}
-	
-	/**
-	 * Convert a map to JSON text. The result is a JSON object. 
-	 * If this map is also a JSONAware, JSONAware specific behaviours will be omitted at this top level.
-	 * 
-	 * @see org.json.simple.JSONValue#toJSONString(Object)
-	 * 
-	 * @param map description omitted.
-	 * @return JSON text, or "null" if map is null.
-	 */
-	public static String toJSONString(Map map){
-		final StringWriter writer = new StringWriter();
-		
-		try {
-			writeJSONString(map, writer);
-			return writer.toString();
-		} catch (IOException e) {
-			// This should never happen with a StringWriter
-			throw new RuntimeException(e);
-		}
-	}
-	
-	public String toJSONString(){
-		return toJSONString(this);
-	}
-	
-	public String toString(){
-		return toJSONString();
-	}
-
-	/**
-	 * description omitted.
-	 *
-	 * @param key description omitted.
-	 * @param value description omitted.
-	 * @return description omitted.
-	 */
-	public static String toString(String key,Object value){
-        StringBuffer sb = new StringBuffer();
-        sb.append('\"');
-        if(key == null)
-            sb.append("null");
-        else
-            JSONValue.escape(key, sb);
-		sb.append('\"').append(':');
-		
-		sb.append(JSONValue.toJSONString(value));
-		
-		return sb.toString();
-	}
-	
-	/**
-	 * Escape quotes, \, /, \r, \n, \b, \f, \t and other control characters (U+0000 through U+001F).
-	 * It's the same as JSONValue.escape() only for compatibility here.
-	 * 
-	 * @see org.json.simple.JSONValue#escape(String)
-	 * 
-	 * @param s description omitted.
-	 * @return description omitted.
-	 */
-	public static String escape(String s){
-		return JSONValue.escape(s);
-	}
-}
